Improving look and feel of home page.

+++ SocialNetwork/SocialNetwork/Models/Concert.cs
@@ -18,10 +18,11 @@ namespace SocialNetwork.Models
 		public string Venue { get; set; }
 
 
+
 		public Genre Genre { get; set; }
 
 		[Required]
-		public byte GenreId { get; set; }
+		public int GenreId { get; set; }
 
 	}
 }
